+++ src/GESRE/excepcion/EmailExisteException.java
@@ -1,20 +1,20 @@
+package GESRE.excepcion;
+
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package GERSE.excepcion;
-
 /**
  *
  * @author Jonathan Viñan
  */
-public class LoginNoExisteException extends Exception {
+public class EmailExisteException extends Exception {
 
     /**
      * Constructor vacío.
      */
-    public LoginNoExisteException() {
-        super("El login que se ha buscado no existe en la base de datos");
+    public EmailExisteException() {
+        super("El email que se ha buscado ya existe en la base de datos");
     }
 }
