+++ src/GESRE/excepcion/EmailNoExisteException.java
@@ -1,20 +1,20 @@
-package GERSE.excepcion;
-
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
+package GESRE.excepcion;
+
 /**
  *
  * @author Jonathan Viñan
  */
-public class EmailExisteException extends Exception {
+public class EmailNoExisteException extends Exception {
 
     /**
      * Constructor vacío.
      */
-    public EmailExisteException() {
-        super("El email que se ha buscado ya existe en la base de datos");
+    public EmailNoExisteException() {
+        super("El email que se ha buscado no existe en la base de datos");
     }
 }
