+++ src/GESRE/excepcion/LoginExisteException.java
@@ -3,18 +3,17 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package GERSE.excepcion;
+package GESRE.excepcion;
 
 /**
  *
  * @author Jonathan Viñan
  */
-public class UsuarioNoExisteException extends Exception {
-
-    /**
+public class LoginExisteException  extends Exception{
+     /**
      * Constructor vacío.
      */
-    public UsuarioNoExisteException() {
-        super("El usuario que se ha buscado no existe en la base de datos");
+    public LoginExisteException() {
+        super("El login que se ha buscado ya existe en la base de datos");
     }
 }
