+++ src/GESRE/excepcion/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Este paquete contiene la clase exception
  */
-package GERSE.excepcion;
+package GESRE.excepcion;
