+++ src/GESRE/seguridad/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Este paquete contiene la clase MAIN
  */
-package GERSE.seguridad;
+package GESRE.seguridad;
